Arreglado problema al guardar minutos de retraso en base de datos, agregado mas funciones en la vista de proyecto de marketing

+++ database/migrations/2022_03_17_112205_create_pay_roll_registers_table.php
@@ -28,7 +28,7 @@ class CreatePayRollRegistersTable extends Migration
             
             $table->boolean('extras_enabled')->default(0);
 
-            $table->unsignedTinyInteger('late')->default(0);
+            $table->unsignedInteger('late')->default(0);
 
             $table->foreignId('pay_roll_id')
                 ->constrained()
